# BACKUP/New-Khaffah-Backend-main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kaffah API",
        "description": "Dokumentasi API Kaffah (frontend & admin)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/api/auth/google": {
            "get": {
                "tags": [
                    "Auth Google"
                ],
                "summary": "Redirect ke Google OAuth",
                "description": "Mengarahkan user ke halaman login Google untuk autentikasi",
                "operationId": "e299709dd3227a1a35cee7cf5f4eb21b",
                "responses": {
                    "302": {
                        "description": "Redirect ke halaman login Google"
                    },
                    "400": {
                        "description": "Error saat redirect"
                    }
                }
            }
        },
        "/api/auth/google/callback": {
            "get": {
                "tags": [
                    "Auth Google"
                ],
                "summary": "Handle callback dari Google OAuth",
                "description": "Memproses callback dari Google setelah user login, membuat atau memperbarui user, dan mengembalikan token authentication",
                "operationId": "cfd1704a13d84d045c2b5a6dc0cad092",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Authorization code dari Google",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "State parameter untuk keamanan CSRF",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect ke frontend dengan token dan data user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer abc123..."
                                        },
                                        "nama": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "foto": {
                                            "type": "string",
                                            "format": "uri",
                                            "example": "https://lh3.googleusercontent.com/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token Google tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token Google tidak valid"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid authorization code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}